+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.4.0-rc6](https://github.com/rusty-celery/rusty-celery/releases/tag/v0.4.0-rc6) - 2020-12-11
+
 ### Added
 
-- Redis broker support.
+- 🚀🚀 Redis broker support 🚀🚀
 - Added the `max_sleep_duration` property on the `Beat` which can be used to ensure that
   the scheduler backend is called regularly (which may be necessary for custom backends).